Added items search
broke out searchs into seperate components

added react-router-dom and useDebounce hook
