+++ package.json
@@ -11,6 +11,7 @@
     "graphql": "^16.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "^5.0.0",
     "web-vitals": "^1.0.1"
   },
